+++ ImageFeed/Profile/ProfileServices/ProfileService.swift
@@ -1,33 +1,5 @@
 import Foundation
 
-struct ProfileResult: Codable {
-    let username: String
-    let firstName: String?
-    let lastName: String?
-    let bio: String?
-    
-    enum CodingKeys: String, CodingKey {
-        case username
-        case firstName = "first_name"
-        case lastName = "last_name"
-        case bio
-    }
-}
-
-struct Profile {
-    let username: String
-    let firstName: String
-    let lastName: String
-    let bio: String
-    
-    var name: String {
-        return "\(firstName) \(lastName)"
-    }
-    var loginName: String {
-        return "@\(username)"
-    }
-}
-
 final class ProfileService {
     static let shared = ProfileService()
     private init() {}
